+++ include/WindowSDL.h
@@ -29,8 +29,6 @@ public:
      */
     virtual ~WindowSDL();
 
-    virtual void setDriver(const char *driver);
-
     virtual void setSize(unsigned int width, unsigned int height);
 
     virtual void setFullscreen(bool fullscreen);
@@ -47,26 +45,12 @@ public:
 
     virtual void swapBuffersGL();
 
-    virtual void setFont(const char *font);
-
-    virtual int initializeFont();
-
-    virtual void writeString(WindowString &s);
-
-    virtual void drawText(unsigned int x, unsigned int y, float scale, const float color[4], const char *s, ...)
-        __attribute__((format(printf, 6, 0)));
-
 private:
-
     SDL_Window *mWindow;      //!< This is the pointer to the SDL surface
     SDL_GLContext mGLContext; //!< The OpenGL Context
 
-    TTF_Font *mFont;
-    unsigned int mFontTexture;
-
     bool mTextInput;
-
-    WindowString tempText;
 };
 
 #endif
+