+++ src/Font.cpp
@@ -1,0 +1,24 @@
+/*!
+ * \file src/Font.cpp
+ * \brief Font implementation
+ *
+ * \author xythobuz
+ */
+
+#include "global.h"
+#include "utils/strings.h"
+#include "Font.h"
+
+Font::~Font() {
+    if (mFontName)
+        delete [] mFontName;
+}
+
+void Font::setFont(const char *font) {
+    assert(font != NULL);
+    assert(font[0] != '\0');
+    assert(mFontInit == false);
+
+    mFontName = fullPath(font, 0);
+}
+