+++ src/Window.cpp
@@ -11,9 +11,28 @@
 
 #include "global.h"
 #include "math/math.h"
+#include "utils/strings.h"
 #include "Window.h"
 
 Window::~Window() {
+    if (mDriver)
+        delete [] mDriver;
+}
+
+unsigned int Window::getWidth() {
+    return mWidth;
+}
+
+unsigned int Window::getHeight() {
+    return mHeight;
+}
+
+void Window::setDriver(const char *driver) {
+    assert(driver != NULL);
+    assert(driver[0] != '\0');
+    assert(mInit == false);
+
+    mDriver = bufferString("%s", driver);
 }
 
 int Window::initializeGL() {